Make QADP capacity functions use qsizetype instead of size_t
Change types returned and accepted by capacity-related QArrayDataPointer
functions to qsizetype:

1) QArrayData (underlying d-ptr) works with qsizetype
2) QArrayDataPointer::size is of type qsizetype
3) All higher level classes that use QADP (e.g. containers)
   cast capacity to qsizetype in their methods

Additionally, fixed newly appeared warnings through qtbase

+++ src/corelib/text/qbytearray.cpp
@@ -1173,8 +1173,8 @@ QByteArray &QByteArray::operator=(const char *str)
     } else {
         const qsizetype len = qsizetype(strlen(str));
         const auto capacityAtEnd = d->allocatedCapacity() - d.freeSpaceAtBegin();
-        if (d->needsDetach() || size_t(len) > capacityAtEnd
-                || (len < size() && size_t(len) < (capacityAtEnd >> 1)))
+        if (d->needsDetach() || len > capacityAtEnd
+                || (len < size() && len < (capacityAtEnd >> 1)))
             reallocData(len, d->detachFlags());
         memcpy(d.data(), str, len + 1); // include null terminator
         d.size = len;
